split up the kerchunk tests into tests of writing and reading kerchunk
